move actions, utils, models to /entities subpackage
