+++ packages/components/src/internal/components/entities/utils.spec.ts
@@ -1,19 +1,11 @@
 import { List } from 'immutable';
 
-import mixturesQueryInfo from '../../../test/data/mixtures-getQueryDetails.json';
 
-import { SchemaQuery } from '../../../public/SchemaQuery';
-
-import { makeQueryInfo } from '../../testHelpers';
-
-import { EntityChoice, IEntityTypeOption } from './models';
+import { IEntityTypeOption } from './models';
 import {
-    createEntityParentKey,
     getEntityDescription,
     getEntityNoun,
     getInitialParentChoices,
-    getUpdatedLineageRowsForBulkEdit,
-    parentValuesDiffer,
 } from './utils';
 import { DataClassDataType, SampleTypeDataType } from './constants';
 
@@ -175,275 +167,6 @@ describe('getInitialParentChoices', () => {
     });
 });
 
-describe('parentValuesDiffer', () => {
-    test('empty lists', () => {
-        expect(parentValuesDiffer(List<EntityChoice>(), List<EntityChoice>())).toBe(false);
-    });
-    const original = List<EntityChoice>([
-        {
-            type: {
-                lsid: 'lsid1',
-                rowId: 1,
-                label: 'Label 1',
-            },
-            ids: ['id1', 'id2'],
-            value: 'Val1,Val2',
-        },
-        {
-            type: {
-                lsid: 'lsid2',
-                rowId: 2,
-                label: 'Label 2',
-            },
-            ids: ['id3'],
-            value: 'Val3',
-        },
-    ]);
-
-    test('non-empty without differences', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(false);
-    });
-
-    test('order change only', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(false);
-    });
-
-    test('original list smaller', () => {
-        const original = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-        ]);
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(true);
-    });
-
-    test('original list larger', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(true);
-    });
-
-    test('same size, different types', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid31',
-                    rowId: 31,
-                    label: 'Label 31',
-                },
-                ids: ['id3.1'],
-                value: 'Val3.1',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(true);
-    });
-
-    test('same types, different values', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val2,Val1.1',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(true);
-    });
-
-    test('same types, same values, different order', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val2,Val1',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(false);
-    });
-
-    test('current list with unspecified type', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-            {
-                type: {
-                    lsid: undefined,
-                    label: undefined,
-                    rowId: undefined,
-                },
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(false);
-    });
-
-    test('current list with new type specified but no value', () => {
-        const current = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                },
-                ids: ['id1', 'id2'],
-                value: 'Val1,Val2',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                },
-                ids: ['id3'],
-                value: 'Val3',
-            },
-            {
-                type: {
-                    lsid: undefined,
-                    label: 'Label 4',
-                    rowId: 4,
-                },
-                ids: undefined,
-                value: undefined,
-            },
-        ]);
-        expect(parentValuesDiffer(original, current)).toBe(false);
-    });
-});
-
-describe('createEntityParentKey', () => {
-    test('without id', () => {
-        expect(createEntityParentKey(SchemaQuery.create('schema', 'query'))).toBe('schema:query');
-    });
-    test('with id', () => {
-        expect(createEntityParentKey(SchemaQuery.create('schema', 'query'), 'id')).toBe('schema:query:id');
-    });
-});
-
 describe('getEntityDescription', () => {
     test('zero', () => {
         expect(getEntityDescription(DataClassDataType, 0)).toBe('parent types');
@@ -465,257 +188,5 @@ describe('getEntityNoun', () => {
     });
     test('multiple', () => {
         expect(getEntityNoun(SampleTypeDataType, 2)).toBe('samples');
-    });
-});
-
-describe('getUpdatedLineageRowsForBulkEdit', () => {
-    const QUERY_INFO = makeQueryInfo(mixturesQueryInfo);
-    const original1 = List<EntityChoice>([
-        {
-            gridValues: [
-                {
-                    displayValue: 'Val1',
-                },
-                {
-                    displayValue: 'Val2',
-                },
-            ],
-            type: {
-                lsid: 'lsid1',
-                rowId: 1,
-                label: 'Label 1',
-            },
-            ids: ['id1.1', 'id1.2'],
-            value: undefined,
-        },
-        {
-            gridValues: [
-                {
-                    displayValue: 'Val3',
-                },
-            ],
-            type: {
-                lsid: 'lsid2',
-                rowId: 2,
-                label: 'Label 2',
-            },
-            ids: ['id2.1'],
-            value: undefined,
-        },
-    ]);
-    const original2 = List<EntityChoice>([
-        {
-            gridValues: [
-                {
-                    displayValue: 'Val1',
-                },
-            ],
-            type: {
-                lsid: 'lsid1',
-                rowId: 1,
-                label: 'Label 1',
-            },
-            ids: ['id1.1'],
-            value: 'Val1',
-        },
-    ]);
-
-    const selected = List<EntityChoice>([
-        {
-            type: {
-                lsid: 'lsid1',
-                rowId: 1,
-                label: 'Label 1',
-                entityDataType: SampleTypeDataType,
-            },
-            ids: ['id1.1', 'id1.2'],
-            value: 'Val1,Val2',
-        },
-    ]);
-    const samples = {
-        '1': {
-            rowId: {
-                value: '1',
-            },
-        },
-        '2': {
-            rowId: {
-                value: '2',
-            },
-        },
-    };
-    test('no samples', () => {
-        expect(getUpdatedLineageRowsForBulkEdit({}, selected, {}, QUERY_INFO)).toStrictEqual([]);
-    });
-
-    test('no selected parents', () => {
-        expect(
-            getUpdatedLineageRowsForBulkEdit(
-                samples,
-                List<EntityChoice>(),
-                {
-                    '1': original1,
-                },
-                QUERY_INFO
-            )
-        ).toStrictEqual([]);
-    });
-
-    test('no original parents', () => {
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': List<EntityChoice>(), '2': List<EntityChoice>() },
-            QUERY_INFO
-        );
-        expect(rows).toStrictEqual([
-            {
-                RowId: '1',
-                'MaterialInputs/Label 1': 'Val1,Val2',
-            },
-            {
-                RowId: '2',
-                'MaterialInputs/Label 1': 'Val1,Val2',
-            },
-        ]);
-    });
-
-    test('change one original parent', () => {
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': original1, '2': original2 },
-            QUERY_INFO
-        );
-        expect(rows).toStrictEqual([
-            {
-                RowId: '2',
-                'MaterialInputs/Label 1': 'Val1,Val2',
-            },
-        ]);
-    });
-
-    test('remove one of many original parents', () => {
-        const selected = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                    entityDataType: SampleTypeDataType,
-                },
-                ids: [],
-                value: undefined,
-            },
-        ]);
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': original1, '2': original2 },
-            QUERY_INFO
-        );
-        expect(rows).toHaveLength(1);
-        expect(rows).toStrictEqual([
-            {
-                RowId: '1',
-                'MaterialInputs/Label 2': null,
-            },
-        ]);
-    });
-
-    test('remove original parent entirely', () => {
-        const selected = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                    entityDataType: SampleTypeDataType,
-                },
-                ids: ['id1.1'],
-                value: 'Val1',
-            },
-        ]);
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': original1, '2': original2 },
-            QUERY_INFO
-        );
-        expect(rows).toStrictEqual([
-            {
-                RowId: '1',
-                'MaterialInputs/Label 1': 'Val1',
-            },
-        ]);
-    });
-
-    test('multiple updates', () => {
-        const selected = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                    entityDataType: SampleTypeDataType,
-                },
-                ids: ['id1.2'],
-                value: 'Val2',
-            },
-            {
-                type: {
-                    lsid: 'lsid2',
-                    rowId: 2,
-                    label: 'Label 2',
-                    entityDataType: SampleTypeDataType,
-                },
-                ids: [],
-                value: undefined,
-            },
-        ]);
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': original1, '2': original2 },
-            QUERY_INFO
-        );
-        expect(rows).toStrictEqual([
-            {
-                RowId: '1',
-                'MaterialInputs/Label 1': 'Val2',
-                'MaterialInputs/Label 2': null,
-            },
-            {
-                RowId: '2',
-                'MaterialInputs/Label 1': 'Val2',
-            },
-        ]);
-    });
-
-    test('parents same except for ordering', () => {
-        const selected = List<EntityChoice>([
-            {
-                type: {
-                    lsid: 'lsid1',
-                    rowId: 1,
-                    label: 'Label 1',
-                    entityDataType: SampleTypeDataType,
-                },
-                ids: ['id1.2, id1.1'],
-                value: 'Val2,Val1',
-            },
-        ]);
-        const rows = getUpdatedLineageRowsForBulkEdit(
-            samples,
-            selected,
-            { '1': original1, '2': original2 },
-            QUERY_INFO
-        );
-        expect(rows).toStrictEqual([
-            {
-                RowId: '2',
-                'MaterialInputs/Label 1': 'Val1,Val2',
-            },
-        ]);
     });
 });
